# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": " NodeJS API project to demonstrate Swagger ",
        "version": "1.0.0",
        "title": "E-commers Project API"
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:7065",
    "basePath": "/api/v1",
    "paths": {
        "/register/user": {
            "post": {
                "summary": "Register User",
                "description": "Create the User ",
                "produces": [
                    "application/json"
                ],
                "requestBody": true,
                "parameters": [{
                        "in": "body",
                        "name": "body",
                        "description": "User",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "userName": {
                                    "type": "string",
                                    "example": "Aditi"
                                },
                                "email": {
                                    "type": "String",
                                    "example": "aditi@gamil.com"
                                },
    
                                "password": {
                                    "type": "String",
                                    "example": "dhjizs1"
                                },
    
                                "address": {
                                    "type": "Number",
                                    "example": "kanpur"
                                }
                            }
                        }
                }],
                "responses": {
                    "201": {
                        "description": "successful operation ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/registerResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/login/user": {
            "post": {
                "summary": "Login User",
                "description": "login ",
                "produces": [
                    "application/json"
                ],
                "requestBody": true,
                "parameters": [{
                        "in": "body",
                        "name": "body",
                        "description": "User",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "String",
                                    "example": "aditi@gamil.com"
                                },
                                "password": {
                                    "type": "String",
                                    "example": "aditi2459s"
                                }
                            }
                        }
                }],
                "responses": {
                    "201": {
                        "description": "successful operation ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/get/user/{user_Id}": {
            "get": {
                "summary": "Get user",
                "description": "Get user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_Id",
                        "in": "path",
                        "description": "user_id ",
                        "required": true,
                        "type": "string"
                    }, 
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/registerResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/update/user/{user_Id}": {
            "put": {
                "summary": "Update user",
                "description": "Update user ",
                "produces": [
                    "application/json"
                ], 
                "consumes": ["application/json"],
                "requestBody": true,
                "parameters": [
                    {
                        "in": "path",
                        "name": "user_Id",
                        "description": "user_id",
                        "required": true,
                        "type": "string"
                     
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "updateData",
                        "in": "body",
                        "description": " body ",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "userName": {
                                    "type": "string",
                                    "example": "mobile"
                                },
                                "email": {
                                    "type": "String",
                                    "example": "Good"
                                },

                                "address": {
                                    "type": "String",
                                    "example": "tab"
                                }
                            }                        
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/registerResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/delete/user/{user_Id}": {
            "delete": {
                "summary": "Delete user",
                "description": "Delete user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_Id",
                        "in": "path",
                        "description": "user_id ",
                        "required": true,
                        "type": "string"
                    },
                    
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/deleteResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "registerResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "true"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "data": {
                    "type": "object",
                   
                        "properties": {
                            "profile": {
                                "type": "string",
                                "example": "customer"
                            },
                            "userName": {
                                "type": "string",
                                "example": "Aditi"
                            },
                            "email": {
                                "type": "String",
                                "example": "aditi@gamil.com"
                            },

                            "password": {
                                "type": "String",
                                "example": "dhjizs1"
                            },

                            "address": {
                                "type": "Number",
                                "example": "kanpur"
                            },
                            "createdAt": { "type": "string", "format": "date-time", "example": "2023-11-30T07:52:22.910Z" },
                            "updatedAt": { "type": "string", "format": "date-time", "example": "2023-11-30T07:52:22.910Z" }
                        }
                
                    
                   }
            }
        },
        "loginResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "true"
                },
                "message": {
                    "type": "string",
                    "example": "Author login successfully"
                },
                "userId": {
                    "type": "string",
                    "example": "658bccb4f94d9657c973c8e4"
                },
                "data": {
                    "type": "object",
                   
                        "properties": {
                            "token": {
                                "type": "string",
                                "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhdXRob3JJZCI6IjY1ODNjZjk1YzYxMzE5NDk2ZWFhYjQ4NiIsImlhdCI6MTcwMzEzODY2Mn0.Eg7AYd3U-O7a1Gb9dt2LrQeOqbp4JS0UvQwVPrjC7Fk"
                            }
                        }
                
                    
                   }
            }
        },
...
        "deleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted successfully"
                },
                "Session": {
                    "type": "object",
                    "properties": {
                        "acknowledged": {
                            "type": "string",
                            "example": "true"
                        },
                        "deletedCount": {
                            "type": "number",
                            "example": "2"
                        }
                    }
                }
            }
        },
...
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "Number",
                    "example": "400"
                   
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}